Refactor service interfaces and update imports
Moved service interfaces from the 'implement' package to a new 'service' package and updated all relevant imports in controllers and service implementations. Also removed unused DummyController and LoggingController classes to clean up the codebase.
