--- berniestore/src/main/java/com/bernie/berniestore/controller/DummyController.java
@@ -1,58 +1,0 @@
-package com.bernie.berniestore.controller;
-
-import com.bernie.berniestore.dto.UserDTO;
-import jakarta.validation.constraints.Size;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.RequestEntity;
-import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-import java.util.Map;
-
-@RestController
-@RequestMapping("/api/v1/dummy")
-@Validated
-public class DummyController {
-
-    @PostMapping("/create-user")
-    public String createdUser(@RequestBody UserDTO userDTO) {
-        System.out.println(userDTO);
-        return "User created successfully!";
-    }
-
-    @PostMapping("/request-entity")
-    public String createUserWithEntity(RequestEntity<UserDTO> requestEntity){
-        HttpHeaders header = requestEntity.getHeaders();
-        UserDTO userDTO = requestEntity.getBody();
-        return "User created successfully";
-    }
-
-    @GetMapping("/headers")
-    public String readHeaders(@RequestHeader HttpHeaders headers){
-        List<String> location = headers.get("User-Location");
-        return "Received headers with value : " + headers.toString();
-    }
-
-    @GetMapping("/search")
-    public String searchUser(@Size(min = 5, max = 30) @RequestParam(required = false, defaultValue = "Guest", name = "name") String userName){
-        return "Searching for user: " + userName;
-    }
-
-    @GetMapping("/multiple-search")
-    public String multipleSearch(@RequestParam Map<String, String> params) {
-        return "Searching for user :" + params.get("firstName") + " " + params.get("lastName");
-    }
-
-    @GetMapping({"/user/{userId}/posts/{postId}", "/user/{userId}"})
-    public String getUser(@PathVariable(name = "userId") String id,
-                          @PathVariable(required = false) String postId) {
-        return "Searching for user " + id + " and post: " + postId;
-    }
-
-    @GetMapping({"/user/map/{userId}/posts/{postId}", "/user/map/{userId}"})
-    public String getUserUsingMap(@PathVariable Map<String,String> pathVariables) {
-        return "Searching for user : " + pathVariables.get("userId") + " and post : "
-                + pathVariables.get("postId");
-    }
-}
