+++ berniestore/src/main/java/com/bernie/berniestore/service/IContactService.java
@@ -1,4 +1,4 @@
-package com.bernie.berniestore.service.implement;
+package com.bernie.berniestore.service;
 
 import com.bernie.berniestore.dto.ContactRequestDTO;
 import com.bernie.berniestore.dto.ContactResponseDTO;
