+++ berniestore/src/main/java/com/bernie/berniestore/service/IProductService.java
@@ -1,4 +1,4 @@
-package com.bernie.berniestore.service.implement;
+package com.bernie.berniestore.service;
 
 import com.bernie.berniestore.dto.ProductDTO;
